+++ yash-builtin/src/trap/syntax.rs
@@ -128,8 +128,9 @@ fn is_non_negative_integer(s: &str) -> bool {
 
 #[cfg(test)]
 mod tests {
+    use super::super::CondSpec;
     use super::*;
-    use yash_env::trap::{OldCondition, Signal};
+    use yash_env::signal::Name;
 
     #[test]
     fn print_all_for_no_operands() {
@@ -144,7 +145,7 @@ mod tests {
             result,
             Ok(Command::SetAction {
                 action: Action::Default,
-                conditions: vec![(OldCondition::Signal(Signal::SIGINT), Field::dummy("INT"))]
+                conditions: vec![(CondSpec::SignalName(Name::Int), Field::dummy("INT"))]
             })
         );
     }
@@ -156,7 +157,7 @@ mod tests {
             result,
             Ok(Command::SetAction {
                 action: Action::Ignore,
-                conditions: vec![(OldCondition::Signal(Signal::SIGINT), Field::dummy("INT"))]
+                conditions: vec![(CondSpec::SignalName(Name::Int), Field::dummy("INT"))]
             })
         );
     }
@@ -168,7 +169,7 @@ mod tests {
             result,
             Ok(Command::SetAction {
                 action: Action::Command("echo".into()),
-                conditions: vec![(OldCondition::Signal(Signal::SIGINT), Field::dummy("INT"))]
+                conditions: vec![(CondSpec::SignalName(Name::Int), Field::dummy("INT"))]
             })
         );
     }
@@ -181,9 +182,9 @@ mod tests {
             Ok(Command::SetAction {
                 action: Action::Default,
                 conditions: vec![
-                    (OldCondition::Signal(Signal::SIGHUP), Field::dummy("HUP")),
-                    (OldCondition::Signal(Signal::SIGINT), Field::dummy("2")),
-                    (OldCondition::Signal(Signal::SIGTERM), Field::dummy("TERM")),
+                    (CondSpec::SignalName(Name::Hup), Field::dummy("HUP")),
+                    (CondSpec::Number(2), Field::dummy("2")),
+                    (CondSpec::SignalName(Name::Term), Field::dummy("TERM")),
                 ]
             })
         );
@@ -196,7 +197,7 @@ mod tests {
             result,
             Ok(Command::SetAction {
                 action: Action::Ignore,
-                conditions: vec![(OldCondition::Signal(Signal::SIGHUP), Field::dummy("HUP"))]
+                conditions: vec![(CondSpec::SignalName(Name::Hup), Field::dummy("HUP"))]
             })
         );
 
@@ -205,7 +206,7 @@ mod tests {
             result,
             Ok(Command::SetAction {
                 action: Action::Ignore,
-                conditions: vec![(OldCondition::Signal(Signal::SIGQUIT), Field::dummy("QUIT"))]
+                conditions: vec![(CondSpec::SignalName(Name::Quit), Field::dummy("QUIT"))]
             })
         );
     }
@@ -217,7 +218,7 @@ mod tests {
             result,
             Ok(Command::SetAction {
                 action: Action::Default,
-                conditions: vec![(OldCondition::Signal(Signal::SIGHUP), Field::dummy("1"))]
+                conditions: vec![(CondSpec::Number(1), Field::dummy("1"))]
             })
         );
     }
@@ -229,7 +230,7 @@ mod tests {
             result,
             Ok(Command::SetAction {
                 action: Action::Default,
-                conditions: vec![(OldCondition::Exit, Field::dummy("EXIT"))]
+                conditions: vec![(CondSpec::Exit, Field::dummy("EXIT"))]
             })
         );
     }
@@ -241,19 +242,19 @@ mod tests {
             result,
             Ok(Command::SetAction {
                 action: Action::Default,
-                conditions: vec![(OldCondition::Exit, Field::dummy("0"))]
+                conditions: vec![(CondSpec::Number(0), Field::dummy("0"))]
             })
         );
     }
 
     #[test]
     fn action_with_unknown_conditions() {
-        let result = interpret(vec![], Field::dummies(["-", "FOOBAR", "INT", "999999999"]));
+        let result = interpret(vec![], Field::dummies(["-", "FOOBAR", "INT", "9999999999"]));
         assert_eq!(
             result,
             Err(vec![
                 Error::UnknownCondition(Field::dummy("FOOBAR")),
-                Error::UnknownCondition(Field::dummy("999999999")),
+                Error::UnknownCondition(Field::dummy("9999999999")),
             ])
         );
     }
@@ -265,7 +266,7 @@ mod tests {
             result,
             Ok(Command::SetAction {
                 action: Action::Default,
-                conditions: vec![(OldCondition::Signal(Signal::SIGHUP), Field::dummy("1"))]
+                conditions: vec![(CondSpec::Number(1), Field::dummy("1"))]
             })
         );
     }
@@ -277,7 +278,7 @@ mod tests {
             result,
             Ok(Command::SetAction {
                 action: Action::Default,
-                conditions: vec![(OldCondition::Exit, Field::dummy("0"))]
+                conditions: vec![(CondSpec::Number(0), Field::dummy("0"))]
             })
         );
     }
@@ -289,7 +290,7 @@ mod tests {
             result,
             Ok(Command::SetAction {
                 action: Action::Command("-1".into()),
-                conditions: vec![(OldCondition::Exit, Field::dummy("0"))]
+                conditions: vec![(CondSpec::Number(0), Field::dummy("0"))]
             })
         );
     }